requirements.txt
imutils==0.5.3
numpy==1.17.4
opencv-contrib-python==4.1.2.30
Pillow==6.2.1
six==1.13.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Dropzone==1.5.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
pillowcase==2.0.0
requests==2.21.0
urllib3==1.24.2
Werkzeug==0.15.3
torch==1.3.1
torchvision==0.4.2
uvicorn